+++ ChangeLog.md
@@ -17,6 +17,7 @@ and this project adheres to
 - support for import aliases `(import :java.awt.Color :as :Col)`   
   `(import :java.awt.Color)` is equivalent to `(import :java.awt.Color :as :Color)`
 - support for regex rematching to function `regex/matches?`
+- a *grep* module with grep like search for lines in text
 
 ### Changed